travel-planner/src/agents/budget.py
"""Budget Agent - Calculates costs and budget analysis"""

from swarm import Agent
from src.utils.config import Config


def create_budget_agent():
    """
    Create budget agent
    
    Returns:
        Swarm Agent configured for budget estimation
    """
    
    return Agent(
        name="Budget Agent",
        model=Config.OPENAI_MODEL,
        instructions="""You are a travel budget specialist with knowledge of typical travel costs worldwide.

Your task is to provide a comprehensive budget estimate using your built-in knowledge:

1. **Accommodation** - Estimate total cost for hotels/Airbnb (provide budget/mid/luxury options)
2. **Food** - Daily food budget × number of days
3. **Activities** - Entry fees, tours, attractions
4. **Transportation** - Airport transfers, local transport, between locations

**Cost Breakdown Format:**
```
TOTAL ESTIMATED COST: $X,XXX - $X,XXX

Breakdown:
- Accommodation: $X,XXX (X nights × $XX-XX/night)
- Food: $XXX (X days × $XX/day)
- Activities: $XXX
- Transportation: $XXX

Daily Average: $XXX per day
Budget Status: Within/Over/Under budget
```

Use your knowledge of typical costs for the destination and time period.

**Important**: Compare total against the user's budget range and note if adjustments are needed.

When done, say "TRANSFER_TO_SUPERVISOR" to hand back control.
""",
        functions=[]
    )

